requirements.txt
click==6.7 \
    --hash=sha256:29f99fc6125fbc931b758dc053b3114e55c77a6e4c6c3a2674a2dc986016381d \
    --hash=sha256:f15516df478d5a56180fbf80e68f206010e6d160fc39fa508b65e035fd75130b
Flask==0.12 \
    --hash=sha256:7f03bb2c255452444f7265eddb51601806e5447b6f8a2d50bbc77a654a14c118 \
    --hash=sha256:93e803cdbe326a61ebd5c5d353959397c85f829bec610d59cb635c9f97d7ca8b
itsdangerous==0.24 \
    --hash=sha256:cbb3fcf8d3e33df861709ecaf89d9e6629cff0a217bc2848f1b41cd30d360519
Jinja2==2.9.5 \
    --hash=sha256:a7b7438120dbe76a8e735ef7eba6048eaf4e0b7dbc530e100812f8ec462a4d50 \
    --hash=sha256:702a24d992f856fa8d5a7a36db6128198d0c21e1da34448ca236c42e92384825
MarkupSafe==1.0 \
    --hash=sha256:a6be69091dac236ea9c6bc7d012beab42010fa914c459791d627dad4910eb665
Werkzeug==0.12.1 \
    --hash=sha256:918322630c58fa52b89e14c45de3ad78514e18928dac726060bb89cd647f6b83 \
    --hash=sha256:6716830febe9808bb7521fd26db3b398450cbed0886b2b4bea678b87340f534e
gunicorn==19.7.0 \
    --hash=sha256:7b48e797990b7972862bf93c7eec637efb391217485920e6fa5f509376d94727 \
    --hash=sha256:e6a9f2f3c4c7db48bf8eaec14d407ece2c93b7e905eaacec263055c541a4fe17
requests==2.13.0 \
    --hash=sha256:1a720e8862a41aa22e339373b526f508ef0c8988baf48b84d3fc891a8e237efb \
    --hash=sha256:5722cd09762faa01276230270ff16af7acf7c5c45d623868d9ba116f15791ce8
python-decouple==3.0 \
    --hash=sha256:99834c9ff7ce5c5b2f4a18bc0880753e54ea3aaabc3cfb16961a92d5046665df
